requirements.txt
python-telegram-bot==20.0a0
Django==4.0.4
python-dotenv==0.20.0